Properties of the action dialog.
New dialog associated to the properties action.
This dialog displays detailed information about
a merge or a split including files, their sizes
and the progress.

Note: the use of this dialog exposes a bug which
forces the current action to be suspended while
a dialog is opened. It must be fixed someday.

Fix the behavior of the labels which are
contained in the merge widget that did not
return to their normal states when using the
clear button.