Xtremsplit algorithm fix.
General fix for the Xtremsplit algorithm. If the
MD5 sum calculation option is set to 'true', the
MD5 sum of each chunk is calculated and put at
the end of the last chunk. All the MD5 sums are
written side by side. When the user will merge
the created files, the MD5 will be check each
time a file will be read. If, for a reason or
another, two MD5 sums differ, the user will be
warned at the end of the merge.