Fix dialog suspending an action.
It is more a workaround than a fix. We now use
the present() method from gtk window instead of
the run() method from gtk dialog.
So we use the main loop to run the dialogs
instead of using one loop for each dialog.

The real bug has been reported here:
https://bugzilla.gnome.org/show_bug.cgi?id=606796