Speed indicator.
Add a speed indicator to the status widget. The
indicator is a simple icon (taken from the gtk+
stock icons) which has a tooltip text that shows
the speed of the current action.

Also add a short line of code to print the traces
of the exception in the terminal.