New assistant action.
Add an action to display an assistant (no done
yet) that will help the user to create a split
or a merge by asking him some questions. For
now, the action does nothing interesting but it
is available in the menubar and the toolbar.

Also, update some labels of action to add
mnemonics if they did not have any before.