Optimize the view change.
To change the view now we just hide and show the
right widget. Actually, we do not need to remove
the widget and the status bar from the window to
the add the new widget and the status bar.

The new code lets the change first of the view as
fast as the previous code but the next changes
will be faster.