Add the split assistant.
This assistant should help the user to create a
split more easily. It asks to the user some
questions and then update the gtk+ interface.

Also add the assistant action to the toolbar
using a menu tool button to be able to choose the
assistant to display. A dialog to choose between
the assistant also comes with it to handle the
possibility that the user can just press the
button.

Create a dummy assistant action to be able to
update the sensitivity of the assistant menu
item like all other actions even if it does
nothing.