Make GNOME Split open a file.
Files now can be opened using GNOME Split. If the
file which is given in the command line does not
exist, the program will be opened without update
of the interface. If the file is a valid chunk
that can be merged, the program will update and
use the merge widget, in the other case, we
will use the split widget.