Rework of bytes <-> types methods.
It is special to the Xtremsplit algorithm. All
methods used to convert types to little endian
and little endian to types have been reworked.
Methods are less complicated to understand and
the Xtremsplit algorithm now support the date
field of the first file header.

Also cosmetic change in the README file and the
launcher of the GNOME menu.