User interface improvements.
Add a border width to the split and merge widgets
to make them look better. Use only one progress
bar for the entire interface. Giving the fact
that the 2 views use a progress bar, and that we
can use only one view at a time, always
using the same progress bar should be enough.