CLI parser improvement.
Improve the CLI parser to be able to open the
right view when only a file was given as argument
in the command line. If the given file is a first
chunk that can be merged, open it with the merge
view, in the other case, the file file will be
opened with the split view.