Speed improvement for IO operations.
Change the way to notify the view from the
progress of an action. Instead of notifying the
view every write operations, it now uses a
scheduled task that notify the view every 250 ms.
After few tests, it seems that the time to split
and merge files has been divided by 2.