Fix last assistant step not starting the action.
Wait for the interface to be fully updated before
starting the action. The action being processed
by a thread sometimes interrupt the main loop of
GTK and did not wait for the interface to be
fully updated. This problem was the origin of the
error message telling that all the fields were
not filled.